# docs/swagger.yaml
components:
  schemas:
    apperrors.AppError:
      properties:
        code:
          type: string
        message:
          type: string
        params:
          additionalProperties: {}
          type: object
      type: object
...
    database.VersionAuthor:
      properties:
        kind:
          description: 'operator: written in SkyMail; template_seed: written by the
            Template seed from the repo.'
          enum:
          - operator
          - template_seed
          type: string
        name:
          description: The name the writer's token carried when the version was written.
            Null when not known.
          type: string
        sub:
          description: 'The Keycloak subject of the token that wrote the version.
            Null when not known: the first versions, made from templates written before
            versions were kept.'
          type: string
      type: object
...
    handlers.Template:
      properties:
        archived_at:
          type: string
        archived_by:
          type: string
        created_at:
          type: string
        drafts:
          description: 'Each operator''s draft in progress, newest first: the newest
            version an operator wrote of the template, while it is unpublished. An
            operator''s earlier drafts are superseded and not listed. A draft whose
            base_version_id is not published_version_id is stale: a newer version
            was published after it was started, and publishing it takes force.'
          items:
            $ref: '#/components/schemas/handlers.TemplateVersionSummary'
          type: array
          uniqueItems: false
        html_content:
          type: string
        id:
          type: string
        key:
          type: string
        main_mode:
          description: 'The Authoring mode of the Main source the template sends:
            its published version''s. Null for a template with no published version.'
          enum:
          - jsx
          - visual
          - html
          type: string
        name:
          type: string
        plain_text_content:
          type: string
        published_version_id:
          type: string
        react_email_content:
          type: string
        subject:
          type: string
        system:
          type: boolean
        updated_at:
          type: string
      type: object
    handlers.TemplateVersion:
      properties:
        author:
          $ref: '#/components/schemas/database.VersionAuthor'
        base_version_id:
          description: The published version a draft started from; null for a template's
            first version.
          type: string
        created_at:
          type: string
        current:
          description: Whether this is the published version the template sends.
          type: boolean
        html_content:
          description: The Main source rendered as HTML, as the browser or the Template
            seed rendered it.
          type: string
        html_source:
          description: The HTML source; null when the version has none.
          type: string
        id:
          type: string
        jsx_source:
          description: The JSX source, React Email code; null when the version has
            none.
          type: string
        main_mode:
          description: 'The Authoring mode whose source is the Main source: its render
            is what the version sends.'
          enum:
          - jsx
          - visual
          - html
          type: string
        plain_text_content:
          description: The Main source rendered as plain text.
          type: string
        published_at:
          description: When the version was published; null for a draft.
          type: string
        requested_subject:
          description: 'The subject the Template seed sent for its version. It can
            differ from subject: until the seed''s conflict rule, the seed keeps the
            subject a template already has. Null on an operator''s version and on
            the first versions, made from templates written before versions were kept.'
          type: string
        seq:
          description: 1, 2, 3… within the template, in the order versions were written.
          type: integer
        subject:
          type: string
        template_id:
          type: string
        visual_source:
          description: The Visual source, the block editor's document; null when the
            version has none.
          type: object
      type: object
    handlers.TemplateVersionSummary:
      properties:
        author:
          $ref: '#/components/schemas/database.VersionAuthor'
        base_version_id:
          description: The published version a draft started from; null for a template's
            first version.
          type: string
        created_at:
          type: string
        current:
          description: Whether this is the published version the template sends.
          type: boolean
        id:
          type: string
        main_mode:
          description: 'The Authoring mode whose source is the Main source: its render
            is what the version sends.'
          enum:
          - jsx
          - visual
          - html
          type: string
        published_at:
          description: When the version was published; null for a draft.
          type: string
        requested_subject:
          description: 'The subject the Template seed sent for its version. It can
            differ from subject: until the seed''s conflict rule, the seed keeps the
            subject a template already has. Null on an operator''s version and on
            the first versions, made from templates written before versions were kept.'
          type: string
        seq:
          description: 1, 2, 3… within the template, in the order versions were written.
          type: integer
        subject:
          type: string
        template_id:
          type: string
      type: object
...
    requests.PublishTemplateVersion:
      properties:
        force:
          description: Publish even though a newer version was published after the
            draft was started, replacing it. The replaced version stays in the history.
          type: boolean
      type: object
    requests.SaveTemplateDraft:
      properties:
        base_version_id:
          description: 'The published version the draft started from: the template''s
            published_version_id when the editor opened it. Null only for a template
            that has no published version.'
          type: string
        html_content:
          description: The Main source rendered as HTML, with its Go template actions
            intact for the mailer to fill per send.
          type: string
        html_source:
          description: 'The HTML source. Left out or null: kept as it is.'
          type: string
        jsx_source:
          description: 'The JSX source, React Email code. Left out or null: kept as
            it is.'
          type: string
        main_mode:
          description: 'The Authoring mode whose source is the Main source: html_content
            and plain_text_content are its render.'
          enum:
          - jsx
          - visual
          - html
          type: string
        plain_text_content:
          description: The Main source rendered as plain text.
          type: string
        subject:
          type: string
        visual_source:
          description: 'The Visual source, the block editor''s document: a JSON object.
            Left out or null: kept as it is.'
          type: object
      required:
      - html_content
      - main_mode
      - plain_text_content
      - subject
      type: object
...
info:
  contact:
    email: hello@enesgenc.dev
    name: Enes Genç
    url: https://enesgenc.dev
  description: This is the API documentation for Skymail.
  title: Skymail
  version: "1.0"
openapi: 3.1.0
paths:
...
  /templates/{id}/drafts:
    post:
      description: |-
        Records an operator's draft of a Mail template: a version that is sent to nobody until it is published. The template row, which is what is sent, does not change. Changing which source is the Main source is a save too: send the new main_mode and the render its source gives.

        The editor renders the Main source; the server stores the render it is given. It checks what it can without rendering: the fields are there and not blank, the Main source's Authoring mode holds a source, the base is a published version of this template, a Visual source is a JSON object, a JSX source has code in it, and the subject, plain text and HTML parse as the mailer's Go templates.

        Each save is a new version; an operator's newest version, while unpublished, is their draft in progress. A save continues it when it started from the same base, or else starts from the base. Sources left out, or null, are kept from the version the save continues, so a save never drops a source. A save that changes nothing records nothing and answers 200 with the version it would have repeated: the draft in progress, or with none the base.
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/requests.SaveTemplateDraft'
                description: The draft
                summary: draft
        description: The draft
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.TemplateVersion'
          description: 'Nothing changed: the version the save would have repeated'
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.TemplateVersion'
          description: The draft, written
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: validation.error, template.invalid_base, template.main_source_missing
            or template.invalid_body (params.field names the part)
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: template.archived
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: Save a draft of a template
      tags:
      - Templates
...
  /templates/{id}/versions/{versionId}/publish:
    post:
      description: |-
        Makes a draft the version the template sends: the draft is marked published and copied onto the template row — subject, HTML, plain text, and its JSX source (or an empty string) as react_email_content — in one transaction. Answers with the template as publishing left it. Publishing the version the template already sends changes nothing and answers the same way.

        A draft is stale when its base_version_id is not the template's published_version_id: someone published after it was started, and publishing it would quietly revert their version. That is refused with 409 template.stale_base, whose params name version_id (the draft), base_version_id (what it started from) and published_version_id (what is sent now), so both can be shown side by side. {"force": true} publishes it anyway; the replaced version stays in the history.
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: Version ID of the draft
        in: path
        name: versionId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/requests.PublishTemplateVersion'
                description: Whether to publish over a newer version
                summary: publish
        description: Whether to publish over a newer version
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.Template'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: validation.error or template.invalid_body
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: template.stale_base, template.not_a_draft or template.archived
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: Publish a draft
      tags:
      - Templates
  /templates/{id}/versions/{versionId}/restore:
    post:
      description: Copies any version of the template — its subject, sources, Main
        source and render — into a new draft by the caller, started from the version
        published now. Nothing that is sent changes; the draft is published like any
        other. Another operator's draft can be restored too. When the copy would repeat
        the caller's draft in progress, or with none the published version, nothing
        is recorded and that version is answered with 200.
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: Version ID
        in: path
        name: versionId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.TemplateVersion'
          description: 'Nothing changed: the version the copy would have repeated'
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.TemplateVersion'
          description: The draft, written
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: template.archived
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: Restore a version as a draft
      tags:
      - Templates
...
servers:
- url: skymail-api.yildizskylab.com/v1
...